requirements.txt
bcrypt==4.1.3
capstone==5.0.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.1.0
charset-normalizer==3.3.2
colorama==0.4.6
colored-traceback==0.3.0
cryptography==42.0.7
distlib==0.3.6
filebytes==0.10.2
filelock==3.9.0
idna==3.7
intervaltree==3.1.0
invoke==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.0
paramiko==3.4.0
platformdirs==2.6.0
plumbum==1.8.3
psutil==5.9.8
pwntools==4.12.0
pycparser==2.22
pyelftools==0.31
Pygments==2.18.0
PyNaCl==1.5.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0
requests==2.32.2
ROPGadget==7.4
ropper==1.13.8
rpyc==6.0.0
six==1.16.0
sortedcontainers==2.4.0
unicorn==2.0.1.post1
unix-ar==0.2.1
urllib3==2.2.1
virtualenv==20.17.1+ds
zstandard==0.22.0
